move api/login.go -> api/auth/login.go
