+++ router/router.go
@@ -32,14 +32,14 @@
 package router
 
 import (
+	"github.com/gin-gonic/gin"
 	"github.com/go-vela/server/api"
+	"github.com/go-vela/server/api/auth"
 	"github.com/go-vela/server/router/middleware"
 	"github.com/go-vela/server/router/middleware/claims"
 	"github.com/go-vela/server/router/middleware/org"
 	"github.com/go-vela/server/router/middleware/repo"
 	"github.com/go-vela/server/router/middleware/user"
-
-	"github.com/gin-gonic/gin"
 )
 
 const (
@@ -66,7 +66,7 @@ func Load(options ...gin.HandlerFunc) *gin.Engine {
 	r.GET("/health", api.Health)
 
 	// Login endpoint
-	r.GET("/login", api.Login)
+	r.GET("/login", auth.Login)
 
 	// Logout endpoint
 	r.GET("/logout", user.Establish(), api.Logout)
